More test cases and altered URL pattern and validator
This reponds to #63 and #66. Added test cases to prove the issue then amended the pattern to attempt to fix. Simply using the
pattern alone was not reliable enough, there were a couple of edge cases that I have handled in the validator.

It may be that this introduces a tiny amount of overhead but the results are more accurate now. One thing to be aware of is that
url.Parse in the std lib will rarely error and doesn't always generate the Host value as you may expect.

Let me know if you want any amends, or if there are other test cases we are not catching.