Feature: validate custom types that are backed by basic types like byte arrays using a custom type validator.
The custom type validator preserves the field's type and passes an empty interface value to the validation function. I needed this to validate a `gocql.UUID` (`type UUID [16]byte`) in a way that considered a UUID with only zeroes in them as invalid. I also added a unit test that shows how to use the feature. One thing you do lose is the ability to control whether the field is required or optional. Try to handle the decision whether the field is valid or not in your custom validator.

All unit tests are passing.