code now passes the test
The issue we have to work with here is that for something like google or foobar.com the url.Parse populates the Path
but leaves the Host blank. The code in place here is working on the basis of checking the basics before we attempt a
regex match. This seems the most sensible approach to me at this stage