Fix for issue panic with ErrorsByField #112
When ErrorsByField was called with a error of type Errors the items of Errors wasn't verified.
I changed to verify the item's type. In case of being Error, the new error is added to the map and when is Erros the method ErrorsByField  is called recursively.
With only these change will be possible to override some errors with the same name and also the error message starts to being is a little confuse.