Fix data race around accessing custom validators
Because I already introduced a BC break, why not add another? The last
build mysteriously failed due to a data race that happens when accessing
the custom validators. You need to use the new `Get` and `Set` methods
on the `CustomTypeTagMap` struct which protects its private validator
type map with a mutex.