Expose the Validator that Failed in the Error
In some cases it is convenient to have access to the validator that failed
the validation in the error response.  This fix exposes the validator that was
used in the error so it can be used downstream.

My case has it returned in a JSON response so that errors can be shown in a
user interface more conveniently:

```
{
  "error": {
    "code": "validation",
    "message": "Field validation failed",
    "details": {
      "pageAddress": {
        "code": "url",
        "key": "pageAddress",
        "message": "name: non-url does not validate as url"
      }
    }
  }
}
```

This commit allows the "code": "url" part to be available to the error handler.