Validator: Adding more concrete base64 validation
The current base64 validator makes a few assumptions about the nature
of the base 64 string as a result of using regex.
By using Go's build in base64 lib we circumvent the regex as well as
adding full functionally to validate both standard base64 strings
and raw base64 strings.
This is an additive string, which allows for full backwards capabilities.