(Enhancement) Validator: Validating contents of []int
Adding a new type to enable the ability to validate what's inside of an
int slice.  This is important for structured API documentation, when the
items in the slice should be known integers. I didn't implement the other
slice types as I don't have a use case for them, but it should be easy
enough to extend. This is not a breaking change.