apply go vet/lint suggestions
This is mostly a tidy up of comments and erro strings.

One additional point to note is that I dropped a test from validator_test. It wasn't doing anything different to the
test directly above it. It might be nice to refactor some of the tests to use sub-tests too. But I figured
that woud lbe a different PR. Finally, the email tests fail for me here, as they do on master. I've fixed that
on my other PR - where I remove the errant fmt.Println