validator: implement ValidateMap
ValidateMap should be used to validate map[string]interface{} against a
validation map in the form map[string]interface{} and the validation map
should be in the same form as the valid tags. The validation map values
can be either a string or a map[string]interface{} for recursive
validation and the map to be validated can contain values of any kind.